Rename feature.generator to feature.distribution
