+++ src/main/java/cofh/cofhworld/feature/parser/SurfaceParser.java
@@ -1,9 +1,9 @@
 package cofh.cofhworld.feature.parser;
 
-import cofh.cofhworld.feature.generator.FeatureBase;
-import cofh.cofhworld.feature.generator.FeatureBase.GenRestriction;
-import cofh.cofhworld.feature.generator.FeatureGenSurface;
-import cofh.cofhworld.feature.generator.FeatureGenTopBlock;
+import cofh.cofhworld.feature.distribution.FeatureBase;
+import cofh.cofhworld.feature.distribution.FeatureBase.GenRestriction;
+import cofh.cofhworld.feature.distribution.FeatureGenSurface;
+import cofh.cofhworld.feature.distribution.FeatureGenTopBlock;
 import cofh.cofhworld.init.FeatureParser;
 import cofh.cofhworld.util.WeightedRandomBlock;
 import cofh.cofhworld.util.numbers.INumberProvider;
