+++ src/main/java/cofh/cofhworld/feature/parser/UniformParser.java
@@ -2,9 +2,9 @@ package cofh.cofhworld.feature.parser;
 
 import cofh.cofhworld.feature.IFeatureGenerator;
 import cofh.cofhworld.feature.IFeatureParser;
-import cofh.cofhworld.feature.generator.FeatureBase;
-import cofh.cofhworld.feature.generator.FeatureBase.GenRestriction;
-import cofh.cofhworld.feature.generator.FeatureGenUniform;
+import cofh.cofhworld.feature.distribution.FeatureBase;
+import cofh.cofhworld.feature.distribution.FeatureBase.GenRestriction;
+import cofh.cofhworld.feature.distribution.FeatureGenUniform;
 import cofh.cofhworld.init.FeatureParser;
 import cofh.cofhworld.util.WeightedRandomBlock;
 import cofh.cofhworld.util.numbers.INumberProvider;
